// src/modules/challenge/dto/search-challenge.dto.ts
export class SearchChallengeDto {
  user_id?: string;
  search?: string;
  challenge_status?: string;
  language?: string;
  country?: string;
  ids?: string;
  ref_id?: string;
  post_category?: string;
  level_value?: string;
  coin_value?: string;
  channel_id?: string;
  game_type?: string;
  add_all_user?: boolean;
  create_next_cycle?: boolean;
}